requirement.txt
matplotlib == 3.4.3
numpy == 1.21.2
pandas == 1.3.2
pyserial == 3.5